feat(auth):validator with route unit testing done

+++ src/app.js
@@ -3,10 +3,12 @@ const sequelize = require('./config/database');
 const { User, Profile, FileInfo } = require('./model');
 const userRoutes = require("./routes/user.routes")
 const express = require('express');
-sequelize.sync({ force: true })
-    .then(() => {
-        console.log('Database & tables created!');
-});
+if (process.env.NODE_ENV !== 'test') {
+    sequelize.sync({ force: true })
+        .then(() => {
+            console.log('Database & tables created!');
+        });
+}
 
 const app = express();
 app.use(express.json());
@@ -18,3 +20,5 @@ const PORT = process.env.PORT || 3000;
 app.listen(PORT, () => {
     console.log(`Server running on port ${PORT}`);
 });
+
+module.exports = app;
